[FE] Fix: 시설 상세페이지 수정
